[Add] 재고 페이지 컴포넌트
